requirements.txt
click==8.1.2
Flask==2.1.1
gitdb==4.0.9
GitPython==3.1.27
itsdangerous==2.1.2
Jinja2==3.1.1
MarkupSafe==2.1.1
smmap==5.0.0
Werkzeug==2.1.1
